# ui/public/doc/api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Бот-помощник по организации работы команды и увеличения ее эффективности.\nПозволяет выполнять определенные действия по расписанию.",
    "title": "PolkovnikBot API.",
    "contact": {
      "name": "Ladygin Sergey",
      "email": "sladygin@updev.ru"
    },
    "version": "1.1"
  },
  "basePath": "/api",
  "paths": {
...
    "/team/{teamId}/sendMessage": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Отправка сообщения команде.",
        "operationId": "MessageSend",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseSuccess"
          },
          "400": {
            "$ref": "#/responses/ResponseError"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
...
    "/team/{teamId}/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Список заданий команды.",
        "operationId": "TaskList",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskList"
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Создание задания для команды.",
        "operationId": "TaskAdd",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Task"
          },
          "400": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
    "/team/{teamId}/tasks/{taskId}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Информация о задании.",
        "operationId": "TaskGet",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TaskId",
            "description": "ID Задачи",
            "name": "taskId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Task"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Удаление задания.",
        "operationId": "TaskDelete",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TaskId",
            "description": "ID Задачи",
            "name": "taskId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseSuccess"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Изменение задания.",
        "operationId": "TaskEdit",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TaskId",
            "description": "ID Задачи",
            "name": "taskId",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Task"
          },
          "400": {
            "$ref": "#/responses/ResponseError"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
    "/team/{teamId}/tasks/{taskId}/history": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "История выполнения задания.",
        "operationId": "TaskHistory",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "example": 30,
            "x-go-name": "Size",
            "description": "Элементов на страницу",
            "name": "size",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "x-go-name": "Page",
            "description": "Номер страницы",
            "name": "page",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "TaskId",
            "description": "ID Задачи",
            "name": "taskId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskHistories"
          },
          "400": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
    "/team/{teamId}/tasks/{taskId}/run": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Запуск задания вручную.",
        "operationId": "TaskRun",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TaskId",
            "description": "ID Задачи",
            "name": "taskId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseSuccess"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Message": {
      "description": "Сообщение",
      "type": "object",
      "properties": {
        "text": {
          "description": "Текст сообщения",
          "type": "string",
          "x-go-name": "Text"
        }
      },
      "x-go-name": "messageRequest",
      "x-go-package": "polkovnik/api"
    },
    "Task": {
      "description": "Задача",
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "check_date_modify": {
          "type": "string",
          "x-go-name": "DateModify"
        },
        "id": {
          "type": "string",
          "x-go-name": "Id"
        },
        "last_run_time": {
          "type": "string",
          "x-go-name": "LastRunTime"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "projects": {
          "type": "string",
          "x-go-name": "Projects"
        },
        "schedule_hour": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ScheduleHour"
        },
        "schedule_minute": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ScheduleMinute"
        },
        "schedule_weekdays": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "ScheduleWeekdays"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-name": "taskResponseItem",
      "x-go-package": "polkovnik/api"
    },
    "TaskHistories": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TaskHistory"
          },
          "x-go-name": "Result"
        },
        "total": {
          "description": "Количество найденных записей всего",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-name": "taskHistoryResponseList",
      "x-go-package": "polkovnik/api"
    },
    "TaskHistory": {
      "description": "История выполнения задания",
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "x-go-name": "Date"
        },
        "error": {
          "type": "boolean",
          "x-go-name": "Error"
        },
        "id": {
          "type": "string",
          "x-go-name": "Id"
        },
        "logs": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Logs"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-name": "taskHistoryResponseItem",
      "x-go-package": "polkovnik/api"
    },
    "Tasks": {
      "description": "Список задач",
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Task"
          },
          "x-go-name": "Result"
        }
      },
      "x-go-name": "taskResponseList",
      "x-go-package": "polkovnik/api"
    },
...
  },
  "responses": {
    "ResponseError": {
      "description": "Стандартное сообщение об ошибке.",
      "headers": {
        "error": {
          "type": "string",
          "description": "Сообщение об ошибке"
        }
      }
    },
    "ResponseSuccess": {
      "description": "Стандартное сообщение об успешном выполнении запроса.",
      "headers": {
        "result": {
          "type": "string",
          "description": "Результат"
        }
      }
    },
    "Task": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Task"
      }
    },
    "TaskHistories": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/TaskHistories"
      }
    },
    "TaskList": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Tasks"
      }
    },
...
  }
}